+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 ## Unreleased
 - Adds [`Objectware.isOwnEmpty`][isOwnEmpty].
+- Adds [`Objectware.has`][has].
 
 [isOwnEmpty]: https://github.com/moll/js-objectware/blob/master/doc/API.md#Objectware.isOwnEmpty
+[has]: https://github.com/moll/js-objectware/blob/master/doc/API.md#Objectware.has
 
 ## 1.4.0 (Feb 8, 2015)
 - Adds [`Objectware.reject`][reject].