+++ CHANGELOG.md
@@ -4,12 +4,14 @@
 - Adds [`Objectware.hasOwn`][hasOwn].
 - Adds [`Objectware.each`][each].
 - Adds [`Objectware.forEach`][forEach].
+- Adds [`Objectware.eachOwn`][eachOwn].
 
 [isOwnEmpty]: https://github.com/moll/js-objectware/blob/master/doc/API.md#Objectware.isOwnEmpty
 [has]: https://github.com/moll/js-objectware/blob/master/doc/API.md#Objectware.has
 [hasOwn]: https://github.com/moll/js-objectware/blob/master/doc/API.md#Objectware.hasOwn
 [each]: https://github.com/moll/js-objectware/blob/master/doc/API.md#Objectware.each
 [forEach]: https://github.com/moll/js-objectware/blob/master/doc/API.md#Objectware.forEach
+[eachOwn]: https://github.com/moll/js-objectware/blob/master/doc/API.md#Objectware.eachOwn
 
 ## 1.4.0 (Feb 8, 2015)
 - Adds [`Objectware.reject`][reject].